yet another insights page restructure (kind of) and fixes to pie chart overlapping nightmare
