+++ packages/frontend/app/(dashboard)/course/[cid]/(insights)/components/InsightCard.tsx
