+++ packages/frontend/app/(dashboard)/course/[cid]/(insights)/components/InsightsPageMenu.tsx
