+++ packages/frontend/app/(dashboard)/course/[cid]/(insights)/insights/chatbot/page.tsx
@@ -1,0 +1,2 @@
+import InsightsPage from '@/app/(dashboard)/course/[cid]/(insights)/insights/page'
+export default InsightsPage
