+++ packages/frontend/app/(dashboard)/course/[cid]/(insights)/insights/tool_usage_statistics/page.tsx
@@ -1,0 +1,2 @@
+import InsightsPage from '@/app/(dashboard)/course/[cid]/(insights)/insights/page'
+export default InsightsPage
