+++ packages/frontend/app/(dashboard)/course/[cid]/(insights)/layout.tsx
